Refactor build script and clean up comments
Simplified the `build_kernel.sh` script by removing unused echoes, prompts, and commented-out lines. Deleted a redundant phase in `pom.xml` and added a comprehensive `configure-help.txt` for OpenJDK configuration options.

+++ java/scripts/build_java.sh
@@ -1,35 +1,24 @@
 #!/bin/sh
 # shellcheck disable=SC2164
 # shellcheck disable=SC2162
-
-# Constants
 # shellcheck disable=SC2034
-#JDK_URL=""
-#JDK_ARCHIVE="openjdk-23.0.1_linux-x64_bin"
-#echo "$PWD"
-
-#BOOT_JDK="/home/rajames/PROJECTS/StarshipOS/jdk-23.0.1"
-
-# Step 1: Download JDK archive if it does not exist
-#if [ ! -f "openjdk-23.0.1_linux-x64_bin.tar.gz" ]; then
+if [ ! -d build ]; then
   echo "Downloading JDK archive..."
   wget --progress=bar "https://download.java.net/java/GA/jdk23.0.1/c28985cbf10d4e648e4004050f8781aa/11/GPL/openjdk-23.0.1_linux-x64_bin.tar.gz"
   echo "Unpacking JDK archive..."
   tar -xzf openjdk-23.0.1_linux-x64_bin.tar.gz
   rm openjdk-23.0.1_linux-x64_bin.tar.gz
   cd ./jdk
-  echo "Configuring build with boot JDK..."
-  echo "$PWD"
-  read -p "Configure build. Press any key to continue."
-  ./configure --with-boot-jdk="/home/rajames/PROJECTS/StarshipOS/java/jdk-23.0.1"
+  echo "Configuring build with boot JDK... $PWD"
+  ./configure --with-boot-jdk="/home/rajames/PROJECTS/StarshipOS/java/jdk-23.0.1" --with-jobs=4 #--with-conf-name="starship-os"
   echo "Cleaning previous builds if any..."
   make clean
   echo "Building JDK..."
-  make images
+  make hotspot
+  make -J 4
 
-  rm -rfv jdk-23.0.1
-#else
-#fi
+  rm -rfv ../jdk-23.0.1
+fi
 
 # Step 2: Unpack the archive if the directory does not exist
 #if [ ! -d "$EXTRACTED_DIR" ]; then
